59: разобрал CommandServiceImpl и создал кучу новых классов(еще нужно дорабатывать)
