+++ src/main/java/pro/sky/animalsheltertelegrambot/service/ReportService.java
@@ -1,5 +1,7 @@
 package pro.sky.animalsheltertelegrambot.service;
 
+import com.pengrad.telegrambot.model.Message;
+import com.pengrad.telegrambot.request.SendMessage;
 import org.springframework.stereotype.Service;
 import pro.sky.animalsheltertelegrambot.model.Report;
 
@@ -9,4 +11,7 @@ public interface ReportService {
     Report getReport(Long id);
     void updateReport(Long id, Report report);
     void deleteReport(Long id);
+
+    public void saveReport(Message message);
+    public SendMessage displayReportInfo(Long chatId);
 }
