+++ src/main/java/pro/sky/animalsheltertelegrambot/telegram_bot/service/CallbackService.java
@@ -1,4 +1,10 @@
 package pro.sky.animalsheltertelegrambot.telegram_bot.service;
 
+import com.pengrad.telegrambot.model.CallbackQuery;
+
 public interface CallbackService {
+
+    public void handleCallback(CallbackQuery callbackQuery);
+
+    public void processCallback(CallbackQuery callbackQuery);
 }
